feat: setting 페이지 프로젝트 정보 수정 API 연결
